feat: move the lifecycle into the loader
